stack/requirements-dev.txt
aws-cdk-lib==2.28.0
constructs==10.1.41